Stable backend version

+++ node.java
@@ -5,7 +5,7 @@ public class node
     private int traffic; //thte traffic intensity will vary dynamically
     private int distance; //the distance is fixed and is expressed in metres
     private int tvertex;  //indicates the terminal vertex
-    private double speed;  //measure of how fast will it take to reach this node
+    private int speed;  //measure of how fast will it take to reach this node
     private node next;
 
     public node(int distance, int traffic, int tvertex) //constructor
@@ -46,7 +46,7 @@ public class node
 	return this.next;
     }
 
-    public double getSpeed()
+    public int getSpeed()
     {
 	return this.distance/this.traffic;
     }
